requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
ollama==0.1.5
pydantic==2.6.1
httpx>=0.25.2,<0.26.0
pyyaml==6.0
python-multipart==0.0.6

